Restructure Dashboard components + minor changes
