+++ pavilion-web/src/org/components/SecondaryButton.js
@@ -2,7 +2,7 @@
 import React from 'react'
 
 // import stylesheets
-import "../../stylesheets/org/SecondaryButton.scss"
+import "../../stylesheets/SecondaryButton.scss"
 
 const SecondaryButton = ({button_copy}) => {
     return (
