+++ pavilion-web/src/org/components/SelfSignUp/SideBar.js
@@ -1,0 +1,18 @@
+// import essential dependencies
+import React from 'react'
+
+// import stylesheets
+import "../../stylesheets/org/SSUSide.scss"
+
+const SideBar = () => {
+    return (
+        <section className="side-bar">
+            <img src="../assets/pav_logo.svg" alt="Pavilion logo" />
+            <img src="#" alt="" />
+            <div className="blue-ellipse" />
+        </section>
+    )
+}
+
+export default SideBar;
+
