+++ pavilion-web/src/org/views/Dashboard.js
@@ -1,6 +1,6 @@
 import React from 'react' 
-import NavDashboard from '../components/DashboardNav'
-import BodyDashboard from '../components/DashboardBody'
+import NavDashboard from '../components/Dashboard/DashboardNav'
+import BodyDashboard from '../components/Dashboard/DashboardBody'
 import "../../stylesheets/org/Dashboard.scss"
 
 const Dashboard = () => {
