+++ pavilion-web/src/stylesheets/PrimaryButton.scss
@@ -1,4 +1,4 @@
-@import "../variables.scss";
+@import "variables.scss";
 
 .primary_button {
     display: flex;
