+++ pavilion-web/src/stylesheets/SecondaryButton.scss
@@ -1,4 +1,4 @@
-@import "../variables.scss";
+@import "variables.scss";
 
 .secondary_button {
     display: flex;
