+++ pavilion-web/src/stylesheets/org/Dashboard.scss
@@ -5,3 +5,79 @@
     display: flex;
     flex-direction: row;
 }
+/* Dash Body */
+
+.dashboard-body {
+    display: flex;
+    justify-content: center;
+    margin: 0;
+    img {
+        position: static;
+        width: 179px;
+        height: 167.14px;
+        left: 0px;
+        top: 0px;
+    }
+    h1 {
+        position: static;
+        width: 744px;
+        height: 56px;
+        left: 0px;
+        top: 0px;
+        /* text-2xl / bold */
+
+        font-family: Rubik;
+        font-style: normal;
+        font-weight: bold;
+        font-size: 40px;
+        line-height: 56px;
+        /* identical to box height, or 140% */
+
+        color: $blue-300;
+    }
+    p {
+        position: static;
+        width: 744px;
+        height: 64px;
+        left: 0px;
+        top: 72px;
+
+        /* text-md / reg */
+
+        font-family: Rubik;
+        font-style: normal;
+        font-weight: normal;
+        font-size: 20px;
+        line-height: 32px;
+        /* or 160% */
+
+        color: $gray-700;
+    }
+}
+
+/* Dashboard Nav */
+
+.side-nav {
+    display: flex;
+    flex-direction: column;
+    background-color: $blue-300;
+    height: 100%;
+    width: 200px;
+    position: absolute;
+    left: 0;
+
+    .logo {
+        margin: 80px 60px 80px 32px;
+    }
+
+    .nav-item {
+        font-family: $text-base-med-font-family;
+        font-size: $text-base-med-font-size;
+        font-weight: $text-base-med-font-weight;
+        line-height: $text-base-med-line-height;
+        text-decoration: none;
+        color: #FFFFFF;
+        height: 72px;
+        padding-left: 32px;
+    }
+}
